Bugfix: start service with startForegroundService when playing
In API 26+, services cannot be started by an application if it is considered in a background state (in this app's case, when the UI is hidden).
As a workaround, startForegroundService shoudl be called instead to make the promise that the service will be in the foreground in the next 5 seconds.