Feature : identical queues when restarting application
When the application is started, it prepares for playing the last loaded queue. Due to the full randomness of the default shuffle mode, queues were never sorted the same way.
This change makes queues sorted in a predictable way so that tracks are sorted in the same order as it was when the application was last closed.