Fix: do not filter accent when primary is black
Due to black and white having near 0 as hue, it was impossible to extract red
tones as accent color when primary is black.
The new algorithm takes that into account: if the primary color has low
saturation of very low or very high lightness, then colors of the same
hue can be picked as accent.