Auto-updates when store content changes
This handles 2 kind of changes : track specific and generic.
Track specific changes (when adding or updating) also notifies the containing album and artist that its content has changed.
Generic changes (when deleting) requests for a requery for the whole MUSIC category.
Future implementations of MediaStoreMusicDao must attempt to guess what track(s) has changed during a generic change so that the corresponding album, artist and playlist can be updated.