Fixed notification flashes.
MusicService now stops itself when the playback state moves to STATE_STOPPED or STATE_NONE, destroying itself unless there are bound clients.
This means that swiping the notification when not playing stops the service. While playback is paused, the service will be kept alive (but may eventually be killed by the system on API 26+ but not immediately).