Bug: crash when restarting the application
This might be caused by a memory leak where the ViewModel is retained due to the media browser subscription being not properly unsubscribed.
To prevent this bug from occurring, we have to wait for the MediaBrowser to be connected before subscribing.
Maybe this will be fixed in a later version of the media library (AndroidX ?)