Bugfix: playlist shift or random crash on application restart
When quitting the application and stopping playback, when restarting the cached application process, play queue is regenerated with a shift.
In most cases, this makes tracks have bad metadata.
In the worst case, the application crashes.
It seems this was due to LongSparseArray not copying keys correctly, introducing a shift in its values.