Version 1.1.0 is released !
While users may not directly see a lot of changes from v1.0.2, this is a complete rewrite of the media browser engine.
A new built-in playlist is available: Favorites. It lists tracks that users have listened to the most.
There is also some user interface improvements : rounded corners, color changes, and 2-letters section indexing.
This release is also supposed to introduce Odeon to Google, so that it makes it compatible with Google Assistant and Android Auto.