Improvements on search feature
Search was not very handy before: it only matched elements by their exact name. Also, when an album and/or and artist was involved, its children were returned instead of their own item.
As an effort to add a search feature in the UI, searches are now more flexible, matching an artist, an album or a track if its name contains the query. Results are also ordered, listing the best ones first.
When building a playing queue from a search query, the preparer retrieves the children of the first result if it is browsable. Otherwise, the full result list is used as playing queue.