Improvements in shuffle order management
When generating a play queue from a media id, the selected track is the first item in the queue even when shuffle mode is enabled.
When recreating a play queue in onPrepare, the generating queue is the same (its first item is the same), and the currently playing item is restored to what it was the last time the player stopped.