Extract common code from "mediasource" to "common" and "common-test"
"common" module is included as a dependency in all modules, and contains definitions used everywhere, such as MediaId, RuntimePermissions and AppDispatchers.
"common-test" module is included as a test dependency in all modules and contains test doubles and utilities for definitions in "common".
Those common modules are also used to share common dependencies.