Draft for SpotifyRemoteSource
This differs slightly from the POC : authentication is performed through the SpotifyAccountsService so it can be mocked. Each function returns a Resource wrapper instead of throwing exceptions.
Functions that return a Paging should be reworked in order to return a Flow instead, so that pagination can be performed with Flow operators.