Reworked MediaRepository implementation to use flows instead of actors
This uses the scanReduce operator to calculate the difference between the old and the new media list when a new one is emitted.
The implementation uses a BroadcastChannel to share the latest media list. When an exception is emitted, it is rethrown to consumers and the broadcast channel should be recreated.