Moved BrowserTree from media to service module
The idea is to make MediaRepository the main entry point for media metadata, so that it could be shared across multiple modules (service and spotify-client).
BrowserTree is only a wrapper around MediaRepository to expose media as a structured hierarchy of MediaItems, and therefore should live in the context of MusicService, while MediaRepository could be promoted to Singleton.