Updated Dagger to v2.25.2
This version introduces support for some Kotlin constructs.
When defining a @Module object the @JvmStatic annotation is no longer needed on @Provides functions, making the syntax less verbose.
Note that @JvmStatic is still required when defining @Provides functions in the companion object of an abstract class that defines @Binds functions.