Created the :devmenu module
This module is only meant to be used by developers and is not added in releases.
This includes a special screen that could be used to list media and data retrieved from Spotify.
There is also a draft of RangeSlider, a variant of SeekBar with 2 thumbs to select a range of value.