Introduced MediaStoreDao
MediaStoreDao is an implementation of MediaDao that no longer delegates to MediaProvider. Instead, it directly accesses the MediaStore database and filesystem.
The idea is to remove MediaProvider as it is only used by MediaDaoImpl.
Moreover MediaProvider is the only dependency of MediaDaoImpl, meaning those 2 classes could be merged ; the result is MediaStoreDao.