Fixed a bug where album artworks were no longer displayed on Android Q
As of Android Q, artworks could no longer be accessed from MediaStore.Audio.Albums.ALBUM_ART (deprecated, and its value is always null).
The new way of loading those artworks is to resolve the content Uri of the associated track or album with ContentResolver.openTypedAssetFile or ContentResolver.loadThumbnail.
This revision introduces a custom Glide ModelLoader that decodes those Uris.