Use internal content uri for album arts instead of custom Glide loader
While loading album artworks worked fine in Odeon, it leaded to errors or crashes in client MediaBrowser applications.
There is an internal Content Uri for albums arts: content://media/external/audio/albumart. This is recognized and accessible by client applications without permission, even on Android 10, so sharing the Uri via a FileProvider is no longer necessary.
The downside is that there is no longer a way to know in advance if an album art exists for a given album.