Expand song titles on 2 lines
Some song titles are pretty long and doesn't fit in one line.
Those titles are now displayed on 2 lines when in portrait.
In landscape, we often have enough space to have titles fit in one line,
while there are not that much space available vertically
so we avoid adding lines.