Use album and artist cards in search results
The same transitions are used to transition from a card to its detail.
This use as sophisticated implementation
GridLayoutManager.SpanSizeLookup that should have laid out items
so that items of different types are not on the same row.
But that doesn't work.
Unit tests for that class are great though, so I save them in this commit.